Fix heat plugin support
During the configuration of the heat plugin:
- the 'auth_version' attribute was not set correctly,
- the 'demo_tempestconf' user was not assigned the 'member' role,
- a network within the demo project with a subnet was not
  created.

This was causing the heat plugin to fail.

The 'auth_version' is now set correctly, and the 'demo_tempestconf'
user is assigned the 'member' role when configuring the heat
plugin. Also, a network within the demo project with a subnet
is created.

Story: 2007464
Task: 39151

Change-Id: I6d6f2247ecb13317767ebbf93a413bcaff386141