Fix doc style issues and fix some typos
Removes yaml as a type of the code-block as the rendered
text was hardly readable.
Removes shell-session type of the code-block in case of showing
cat output - the output wasn't rendered at all.
Removes emphasize-lines as the emphasized lines were too bright
and therefore not readable - I found there are issues with
openstackdocstheme and Sphinx 3 [1] and I didn't find any simple
solutions - fixing just a few highlighted lines is not worth it.
A few typos were fixed as well.

[1] https://bugs.launchpad.net/openstack-doc-tools/+bug/1876343

Trying if rebuilding the doc solves the rendering issue:
https://docs.openstack.org/python-tempestconf/latest/user/usage.html

Change-Id: I83ebfdeb949bd0a6cfa99889cb9fbb56d8bcb218