Fix placement endpoint discovery
Placement's endpoints are not versioned (don't contain /<version>
in the URLs opposite to services such as compute or volume)
and therefore we don't have to call no_port_cut_url or do_get
method *with* top_level set to True whose take care of cutting
the URls after the version (including the version).
python-tempestconf is able to discover the endpoint and as it's
not versioned, we don't need to do any further processing.

Story: 2008679
Task: 41980
Change-Id: Id3c270fa4a0b09c19edf6c0c9dba48a2235d1907