Skip test_create_update_server_with_subnet
Let's skip test_create_update_server_with_subnet from
heat_tempest_plugin because it is constantly failing on a timeout
issue:
'failed to reach UPDATE_COMPLETE status within the required time
(1200 s)'
The test is working locally so the failure is not caused by any
tempestconf's recent change or by a tempest.conf issue in general.
In order to unblock tempestconf's CI, the patch excludes the test
in python-tempestconf-tempest-devstack-admin-plugins job.

Change-Id: If9f9b3be758897f25d16326aa6d0f7c53e110122