Fix extension discovery for object_storage
Python-tempestconf assumed that swift's /info endpoint is always the
same. But it may happen that the endpoint is renamed to /swift/info
in certain situations (e.g.: when swift is on ceph).

This patch fixes the issue by trying to access the /info endpoint
first. When the connection to /info endpoint is unsuccessful then
/swift/info endpoint is used as a fallback.

Story: 43092
Change-Id: I4447e1527eae8ec75acb4fb96047c894db75620d