[Trivial] Fix swift's extension discovery
In case the swift's endpoint returned an exception during an
extension query, get_capabilities method returned a string of
an empty dict, which is always "true" and therefore it confues
a later if statement checking if any extensions were found.
Let's make the method return an object - dict - which will
be either true (if contains any data) or false (if it's empty).

Change-Id: Ie7974c97a4b119e41fba7150bcd2a2553300f694