Add multistore image option
This is an attempt to populate tempestconf with
multistore feature enabled if and only if multiple
stores are available.
In Rocky and Stein, Glance has added the ability to
configure multiple stores as an EXPERIMENTAL feature.
This feature is fully supported from the Train cycle.

Please refer [1] for more information
[1] https://specs.openstack.org/openstack/cinder-specs/specs/ussuri/support-glance-multiple-backend.html

Here are two test-projects exercising the results of
this change

1. With multi-store enabled
https://review.rdoproject.org/r/c/testproject/+/37159
2. With multi-store not enabled
https://review.rdoproject.org/r/c/testproject/+/37160

Signed-off-by: Soniya Vyas <svyas@redhat.com>
Change-Id: I45f9dce14b60e9385600c57ee56b50aba3a4476f