Fix TestImageService's unit tests
The test class had hardcoded DEFAULT_IMAGE_DIR value to /img/
which resulted in unit tests failures when this path existed
because the tests expect a call to makedirs which didn't happen
as the DEFAULT_IMAGE_DIR already existed.
The patch fixes that by passing exist_ok=True to makedirs call
which will result in always calling makedirs, whether img_dir
exists or not.

Change-Id: I47d15747536394006f6b502b3025220a5077e547