Fix cinder-backup logic
When you have several controllers, if the first controller have
cinder-backup service disabled, the pacemaker return it as down, and so
the cinder-backup is set to False, even though it is running in a
different controller.
This patch do a loop to check if the cinder-backup is enabled in other
controllers, if so return it as True.

Change-Id: I3a16ec4f7b45267fcb8f9c6d01f28fe40faf7661