Drop py3.6/7, T, U jobs and fix neutron discovery
python 3.6 and 3.7 are dropped (upstream tooling as e.g.
Devstack doesn't support this those anymore) together with
train and ussuri jobs.

refstack-client jobs are non-voting now until a fix on
refstack-client's side has landed. Also victoria jobs are
marked as non-voting because they require a patch on devstack's
side which isn't ready to be merged yet [1].

manila-tempest-plugin needs to be pinned in Victoria and
Wallaby because the plugin requires newer tempest version
than is available there.

The patch also makes neutron version discovery with a not
top level URL, seems like something changed because
the discovery has been failing.

[1] https://review.opendev.org/c/openstack/devstack/+/844774

Change-Id: Ia3d78e84ad9eee515eb8ca017ea84df1fd160929